+++ Fun.SunUI.Templates/MAUIApp/MAUIApp/AppShell.fs
@@ -18,31 +18,29 @@ let homePage =
                 VerticalOptions LayoutOptions.Center
                 Children [
                     Image'() {
-                        With(fun this -> SemanticProperties.SetDescription(this, "Cute dot net bot waving hi to you!"))
                         Source(FileImageSource'() { File "dotnet_bot.png" })
                         HeightRequest 200.
                         HorizontalOptions LayoutOptions.Center
+                        SematicDescription "Cute dot net bot waving hi to you!"
                     }
                     Label'() {
-                        With(fun this -> SemanticProperties.SetHeadingLevel(this, SemanticHeadingLevel.Level1))
                         Text(count |> AVal.map (sprintf "Count = %d"))
                         FontSize 32
                         HorizontalOptions LayoutOptions.Center
+                        SematicHeadingLevel SemanticHeadingLevel.Level1
                     }
                     Label'() {
-                        With(fun this ->
-                            SemanticProperties.SetDescription(this, "Welcome to dot net Multi platform App UI")
-                            SemanticProperties.SetHeadingLevel(this, SemanticHeadingLevel.Level2)
-                        )
                         Text "Welcome to .NET Multi-platform App UI"
                         FontSize 18
                         HorizontalOptions LayoutOptions.Center
+                        SematicDescription "Welcome to dot net Multi platform App UI"
+                        SematicHeadingLevel SemanticHeadingLevel.Level2
                     }
                     Button'() {
-                        With(fun this -> SemanticProperties.SetHint(this, "Counts the number of times you click"))
                         Text "Click Me"
                         Clicked(fun _ -> count.Publish((+) 1))
                         HorizontalOptions LayoutOptions.Center
+                        SematicHint "Counts the number of times you click"
                     }
                 ]
             }