+++ Fun.SunUI.WPF/Fun.SunUI.WPF/Elements.Extensions.fs
@@ -1,0 +1,51 @@
+﻿[<AutoOpen>]
+module Fun.SunUI.WPF.Extensions
+
+open System.Windows.Controls
+open FSharp.Data.Adaptive
+open Fun.SunUI
+open Fun.SunUI.WPF.DslInternals
+
+
+type UIElementBuilder<'Element when 'Element :> System.Windows.UIElement> with
+
+    [<CustomOperation("GridRow")>]
+    member inline this.GridRow([<InlineIfLambda>] builder: BuildElement<'Element>, x) = this.With(builder, (fun this -> Grid.SetRow(this, x)))
+
+    [<CustomOperation("GridRow")>]
+    member inline this.GridRow([<InlineIfLambda>] builder: BuildElement<'Element>, x) =
+        this.With'(
+            builder,
+            (fun this -> adaptive {
+                let! x = x
+                Grid.SetRow(this, x)
+            })
+        )
+
+
+    [<CustomOperation("GridColumn")>]
+    member inline this.GridColumn([<InlineIfLambda>] builder: BuildElement<'Element>, x) = this.With(builder, (fun this -> Grid.SetColumn(this, x)))
+
+    [<CustomOperation("GridColumn")>]
+    member inline this.GridColmn([<InlineIfLambda>] builder: BuildElement<'Element>, x) =
+        this.With'(
+            builder,
+            (fun this -> adaptive {
+                let! x = x
+                Grid.SetColumn(this, x)
+            })
+        )
+
+
+    [<CustomOperation("GridRowSpan")>]
+    member inline this.GridRowSpan([<InlineIfLambda>] builder: BuildElement<'Element>, x) = this.With(builder, (fun this -> Grid.SetRowSpan(this, x)))
+
+    [<CustomOperation("GridRowSpan")>]
+    member inline this.GridRowSpan([<InlineIfLambda>] builder: BuildElement<'Element>, x) =
+        this.With'(
+            builder,
+            (fun this -> adaptive {
+                let! x = x
+                Grid.SetRowSpan(this, x)
+            })
+        )