+++ Fun.SunUI.Avalonia/Fun.SunUI.Avalonia/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.2] - 2022-10-09
+
+- Upgrade dependency packages
+- Add extensions for GridRow GridColumn GriwRowSpan GridColumnSpan RowDefinitions
+
 ## [0.0.1] - 2022-10-08
 
 * First release 🌞 🚀