+++ Fun.SunUI.TerminalGUI/Fun.SunUI.TerminalGUI/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.0.2] - 2022-10-09
+
+- Upgrade dependency packages
+
 ## [0.0.1] - 2022-10-08
 
 * First release 🌞 🚀