+++ Fun.SunUI.WPF/Fun.SunUI.WPF/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.5] - 2022-10-09
+
+- Upgrade dependency packages
+- Add extensions for GridRow GridColumn GriwRowSpan GridColumnSpan
+
 ## [0.0.4] - 2022-10-07
 
 - Update dependency packages