+++ Fun.SunUI/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.0.6] - 2022-10-09
+
+Improve inline
+
 ## [0.0.5] - 2022-10-08
 
 Add MakeActionPropertyBuilder